Extra fix needed for @vicatcu's PR of @chegewara's work See below
So I merged the pull-request mentioned in #4046, and noticed a regression
compared to the code currently in master:

In my app I use light-sleep, and to do so I have to carefully not
just turnoff bluetooth, but stop and delete my various services before
going to sleep. During this process I would call BLEDevice::deinit(false).
That still works fine.

However, when my app woke from light-sleep and needed to restart bluetooth
it was hanging in registerApp:

```
Starting bluetooth
[D][BLEDevice.cpp:80] createServer(): >> createServer
[D][BLEServer.cpp:291] registerApp(): >> registerApp - 1
[D][FreeRTOS.cpp:164] take(): Semaphore taking: name: RegisterAppEvt (0x3ffdd7b0), owner: <N/A> for registerApp
[D][FreeRTOS.cpp:173] take(): Semaphore taken:  name: RegisterAppEvt (0x3ffdd7b0), owner: registerApp
```

To I poked around a bit, and it seems that in the PR #4050 branch, there
is a change compared to master.  A bit of code that used to be #ifdef
was changed to be #ifndef.  Which has the effect of never setting
initialized to false.  If initialized is left true, my call to BLEDevice::init()
was having no effect.

So I've changed this ifdef back to what master had, and now my app works
nicely and the corruption I was previously seeing in my BLE structures
seems fixed (the original goal of pulling in 4050.